move useIsTransferAllowed out
